Works with all 3 quenching algorithms
The previous version properly built and stacked trees only with simple
quenching but not with root-port or my-port quenching.  The error was
overwriting the port state from child to parent on the root port.  All
I had to do was add a test to see if this cell is the root of the tree.