Added test config file
The current edge list connects cells from the outside in so that most
of the time a cell won’t send it’s DiscoverMsgs until all ports that
will be connected are connected.

This test version connects cells row by row, which means the heuristic
will often be violated.  In spite of that, tree building and stacking
works fine.