Generate edge lists from #rows and #cols
Previously, I made edge lists by hand, even for rectangular geometries.
 That made it a pain to try new configurations.  This commit generates
the edge lists automatically.

There is a kludge for defining which cells are border cells that needs
to be fixed.  The datacenter build fails if I pick a cell with
max_num_phys_ports_per_cell neighbors as a border cell, because it
doesn’t have a free port to use as a border port.