Changed tree building heuristic
We want to send out a cell’s DiscoverMsg only after all its ports have
been connected, especially if we’re doing breadth first 1 hop tree
building.  The problem is that a cell can’t know how many of its ports
will be connected.  Hence, we need a heuristic.

Any cell that has all its physical ports connected can safely send its
DiscoverMsg.  Those that don’t need to decide when they’ve waited long
enough.  I was using the number of received DiscoverMsgs, but the
appropriate value depends on the quenching algorithm

This commit uses the number of tree IDs seen.  It’s easy to get a good
value.  In a grid of nrows by ncells, all the interior cells should
have all their physical ports connected.  That may not be always true
due to cabling errors, so we choose a value 10% less than
(nrows-1)*(ncols-1).  Experiment shows we can go as low as half that
number, but cabling errors should be in the few percent range.

The good news is that even if we get the heuristic slightly wrong, the
trees still get built properly.  It’s only if we make the value really
small do we get incomplete tree building.