Gave up on hope for future animation
The web server receives trace records and builds data structures to
allow showing tree building while it’s in progress.  However, the web
server runs on the same machine as the simulator, so it has to compete
for CPU cycles.  The result is server errors that result in some trace
records not being processed.

This commit only shows the end result of tree building by replaying
from a trace file.  It’s slower than showing from the pre-built data
structures, but it has the slight advantage of being correct.