Cellagent::send_msg takes ownership of the message being sent
By default, I pass references, but there are times when the invoked
method should take ownership of one or more of the arguments.  The
send_msg function in cellagent.rs is such a case.  Having it take
ownership of the message makes sure that I can’t change its contents
following the call.