Moved stacking of 3hop and 2hop trees to process_tree_name
The 3hop tree is stacked on the Base tree, so it needs to wait for that
tree_name to be delivered to the Noc.  The 2hop tree needs the
tree_name for the 3hop tree to be delivered.  Hence, it makes more
sense to invoke the small_tree method from process_tree_name.