Fixed noc.rs handling of deploying
I had a kludge in noc.rs that hid a bug in process_stack_tree_d_msg()
in cellagent.rs.  I was simply counting how many tree names the NOC had
received from the cell agent before deploying NocMaster and NocAgent.
Unfortunately, that heuristic needed to be changed after I started
stacking the 3-hop and 2-hop trees, so I decided to do it right.

When I fixed that bug, I discovered that the NOC never got the tree
name for the Master deploy tree.  The problem was that it’s a 0-hop
tree, e.g., only the root cell is on it, and the test I had for the end
of tree building assumed all trees were at least one hop.

When I fixed that bug, the web server would break when processing
StackTreeDMsgs whenever it saw a receive port of 0.  All other ports
have neighbors, but port 0 does not.  The fix was not to process such
messages in the web server.

I also learned about the ..Default::default() syntax for creating a
struct.  It really simplified a couple of them.  However, all elements
in the struct need to implement the Default trait, but channels do not.
 I decided the clarity was worth wrapping them in a Vec, so I could
implement Default.  The downside is that I needed to write
channel[0].send() instead of channel.send().  That change is why I
didn’t use Option, because it takes a lot more characters to unwrap it.