Added a condition to send_my_discover()
There is a complex condition evaluated in send_my_discover() in
cellagent.rs when deciding whether to send a cell’s DiscoverMsg for
breadth-first discover.  It involves the number of neighbors seen and
the number of trees seen.  If I want to run purely stochastic discover,
I had to change the config file to set both min_hello to 1 and
breadth_first_depth to 0.

I added a test for bread_first_depth == 0 to send_my_discover() so I
don’t have to change min_hello when switching to and from breadth first
discover.