Prepare to make a big change
THIS VERSION DOES NOT WORK.  I am committing it because I’m about to
make a big set of changes to produce a version that does work.  Since
I’m not sure my first try will work, I want to be able to back out any
changes.  Unfortunately, I (stupidly) made some other changes that I
don’t want to lose.  Hence this commit.

The original goal was to parameterize the number of hops in the breadth
first Discover process.  That doesn’t produce nice trees because the
first cell outside the breadth first zone grabs the rest of the data
center by immediately forwarding DiscoverMsgs while those in the
breadth first zone are waiting for permission.  Hence, this version is
either breadth first or stochastic.

This version does not work because earlier versions assumed that the
first received DiscoverMsg for a tree determines the parent port.  To
do breadth first Discover, the first received Prepare message
determines the parent.  That wouldn’t be a problem, except updating the
traph data structure assumes that the tree parent has already been
defined when receiving a subsequent DiscoverMsg.  Undoing that
assumption looks really hard.

The changes I want to save are:

1. Distinguish between the originator of a message and the immediate
forwarder for hop-by-hop messages, such as DiscoverMsg.

2. Changed trace records to show the target cell names instead of out
ports, which makes analyzing the traces simpler.