Breadth first tree building working
but tree stacking is not because there’s at least one place where I’m
not updating the hop count in a DiscoverMsg when I should.  I also have
to add in if-tests so it will work with stochastic Discover.  I’m going
to commit anyway so I don’t lose the changes that work.

The big changes are in cellagent.rs.  In my previous commit that didn’t
work, I used the first PrepareMsg I saw for a tree to decide which port
was the parent.  That would have worked, except I had assumed in
traph.rs that the parent was determined by the first time I saw a
DiscoverMsg on that tree.  Going back and changing that assumption was
the right thing to do but way too much work.

This version changes the order of DiscoverMsg, PrepareMsg, and
PrepareDMsg, making the names of the last two misleading.  I’ll fix
that problem in a later commit.

On receiving HelloMsg, the cell sends a DiscoverMsg on that port.  If a
received DiscoverMsg is the first time this tree has been seen, that
port becomes the parent.  The receiving cell sends a PrepareDMsg to the
parent.  Once the parent has heard from all of its neighbors, it sends
a PrepareMsg to all of them.  On receiving a PrepareMsg on a tree, the
cell forwards the DiscoverMsg for that tree (supposedly with updated
hop count) to all its neighbors.

The changes to Cargo.toml allow me to compile my dependent crates with
high optimization and my own files with fast compilation.  I also added
a TODO to utility::Path.