Change Path::new to take PortNumber instead of PortNo
Every place I create a Path, I already know the PortNumber.  In some
cases, I even had to convert a PortNumber to a PortNo to create a Path.
 So, I changed Path to take PortNumber instead of PortNo and the number
of physical ports.  The fact that I no longer return a Result would be
a big deal except that Rust makes handling Result really easy.