DOES NOT WORK: Fixed heuristic in discoverd_done()
I was comparing the number of received HelloMsgs to the number of ports
seen on the tree, which led me to sent DiscoverDMsg::Parent too soon.
That’s not the heuristic, so I now compare with CONFIG.min_hello and
CONFIG.min_trees.  The result is better but still not correct.

I’m committing this non-working version to preserve some improvements I
made to the trace files.