Added back ability to wait for ports to be connected
The tree building heuristic fails when a cell receives a HelloMsg after
it has sent out its DiscoverMsg.  When that happens, it is possible for
a physical neighbor to be more than one hop away on the cell’s tree.
This problem is mitigated by setting Config.min_trees to be large
enough, but the problem can still occur on rare circumstances.

Before I came up with the Config.min_trees idea, I had the cell sleep
before sending its DiscoverMsg based on the value of Config.race_sleep.
 I took out the actual sleep when I implemented Config.min_trees.  This
commit puts that sleep back in.

Experiments simulating a 30-cell cluster Config.race_sleep = 0 shows
the heuristic fails about 10% of the time, although all trees were
correctly formed.  The heuristic was never violated when sleeping for 1
second in the limited number of tests run.