This fixes only a few of the bugs introduced in a recent merge.
Some additions are made to config files.
A change was made to cell.rs to get it to compile.

nrows & ncols are still required in the config file even if num_cells > 0 (3 cell config and previously 10 cell config).
edge_list is still required in the config file even if nrows & ncols are specified (10 cell config now).

cell binary still fails on 10 cell config with "No such file or directory"