Gave NalCell access to all the communication channels
My plan is to read the replay trace records in NalCell and have it send
the appropriate message to the desired endpoint.  To do that, it needs
access to all the senders and receivers.  Unfortunately, the NalCell
starts the cell agent; the cell agent starts the cmodel; and the cmodel
starts the packet engine.  Each step passes the needed communication
channels for that pair.

This commit moves all the channel creation to NalCell, which passes
them to the cell agent, which passes the appropriate ones to the
cmodel.  The cmodel was already passing the right ones to the packet
engine.