Successfully replay messages that set forwarding table entries - DOES NOT RUN TO COMPLETION
NalCell reads all trace records for the cell and forwards
CaToCmBytes::Entry messages to the packet engine via the cmodel.  I
have verified that these messages are identical on the original run and
the replay.

This commit does not run to completion for reasons unknown.  The most
likely problem is that messages from other cells are arriving in a
different order in the replay than in the original run, resulting in an
inconsistency.