Tree building during replay works
I can now read trace records and replay them well enough that tree
building, stacking, deploying and messaging to and from the root works.
 However, there is a problem getting the border port to forward the
messages from the border cell to the NOC.

Still some digging to do.  The forwarding table routing entries are
correct but they don’t appear in the replay and original trace files in
the same order.