Replay working (most of the time)
Fixed a problem with messages to and from the NOC during replay.  I use
the NOC’s OriginatorID to know which trees it is allowed to name.
However, the UUID part of the OriginatorID changes on each run.  I got
around that problem by making that part of OriginatorIDs
Uuid::default().

The name part of the OriginatorID is unique in all the cells of the
simulator but only within a cell on the physical system.  That could
lead to confusion, but there is enough other information in the trace
records that a person should be able to disambiguation any conflicts.

It appears that trees are identical between replay runs (further
testing needed), which is one of the goals.  However, replay does not
guarantee that messages are processed in the same order every time.  In
particular, on some runs messages from the NocAgents reach the
NocMaster before the message adding an AllowedTree gets processed.  I
don’t know if this behavior represents an irreducible problem or not.