Make the replay trace closer to the one it replays from
I thought I had put all message sends behind a test for replay, but
there were some I missed.  The result was many more messages in the
trace file than in the file it was replayed from.

I also fixed the way I handle race_sleep in process_hello_msg() in
cellagent.rs.  The idea is to sleep the cellagent before sending its
DiscoverMsg and DiscoverDMsg in the hope that more ports will be
connected.  In a normal run I put that sleep in a thread, but that
results in a very different message order when replaying.  I now skip
the sleep when replaying.