Works with VSCode
After the latest reorganization of the repository, the IntelliJ editor
started reporting a gazillion spurious errors.  I know they’re spurious
because the program compiles and runs correctly.  I’ve been wanting to
move to VSCode, anyway, so this was a good excuse.

The only files in this commit that really matter are launch.json and
tasks.json.  The others are showing up for reasons I don’t understand.