Added cell_id to port state for trace records
The Port uses the cell_id to construct the name part of the port_id but
did not remember the cell_id after new().  The result was trace records
for all ports appearing in the trace file with Other, such as the
blueprint trace record.  This change adds cell_id to the port state so
it can be used in trace records.