Application replay works
I spent a lot of time tracking down a race condition.  I knew it was a
race condition because application messages appeared out of order on
the console about half the time.

It wasn’t a race condition.  It was the NOC, which gets connected to a
cell based on the order in which border cells appear in a hash map, an
order that can change on each run.  Since the case I was testing only
has two border cells, about half the time the NOC connected to a
different cell than in the replay trace file.

The fix is to have rack.rs create a trace record in a trace file
specific to the rack that gives the cell number used for the NOC.  On
replay, rack.rs reads that cell number and used it to connect to the
NOC.

There’s also some minor cleanup to get rid of some warnings.