Got rid of warning messages
We have tolerated a ton of warning messages.  It isn’t a problem when
they show up on the console, and it was just a minor inconvenience the
way they were handled by IntelliJ.  VSCode is a different matter.  The
warnings are super annoying.

I made a bunch of changes to get rid of them by making more use of
#[cfg…].  In particular, the modules ecnl and ecnl_port are no longer
included when building the simulator.   That change required a bit of a
hack in nalcell.rs and port.rs.  I had to create type aliases for
ECNL_Port and ECNL_Session so they’d show up as known types.