Clear saved packet(s) when hearing from the other side
When a packet is sent on a link, a counter is incremented and a copy of
the packet is kept in case the link loses the packet in transit.  When
a packet is received from the other side, I was only decrementing the
counter.  I should have also been forgetting the packet, which this
commit does.