Add configuration to the trace
We should be capturing the configuration used for a run, but we
weren’t.  It’s now part of the trace.

The config file is read in a lazy_static block in config.rs.  I tried
every which way to get the pub static ref CONFIG into the trace file,
but it isn’t Serialize.  I couldn’t even print it because it isn’t
Display or Debug.  I ended up just reading the file again in
datacenter.rs.  It’s no big deal, since it only happens on startup.