Fixed a problem with flow control
Each port has a buffer of 8 element so it can hold one packet from each
of the other 7 ports and one from the cell agent.  A cell receiving a
packet doesn’t send a pong until it can move the received packet to all
the destination out port buffers.  The problem was how the in port
knows when there’s an available out port slot.

The fix for this problem was to make the outport buffer twice as large.
 If a packet from the in port goes into the first half, the in port
sends its pong.  If it goes into the second half, it doesn’t send the
pong.  When a packet from the top half reaches the bottom half, the
pong for the corresponding in port can be sent.  That means we have to
remember the in port.

I distinguish packets in the first half from those in the second half
using Option<PortNo> for the remembered port number.  Unfortunately, I
forgot to set Some(port_no) to None when a packet moved from the back
half to the front half.  This commit fixes that problem.  The relevant
changes are around line 400 in packet_engine.rs.

This commit also includes some incomplete updates to deal with node
failure.  They can be ignored when reviewing this commit.