Snake data structure created
When the packet engine receives a packet, it creates a snake for that
packet and inserts the packet, the receiving port, and a count of 0
into a hash map with a key of the packet’s uniquifier.  When that
packet gets sent, the count is updated to be the number of out ports.