Accidentally put a send in the wrong brackets
I have a convention of putting trace and debug blocks inside a set of
bare {}.  That makes it easy for me to collapse them in the editor, and
it prevents accidental shadowing of variables.  I have a loop with one
of these blocks and a single line to execute.  Unfortunately, I put the
line to execute in the bare {}.  The program still ran correctly, but
when I collapsed the trace block, the loop looked empty.