I thought I had a bad bug, but it was only a display problem
A packet knows if it’s the last packet.  If it is, it knows how many
data bytes it holds.  My trace files showed that this value was always
the payload size.  I thought that meant that my message assembly was
using too many bytes out of the last packet.  It wasn’t.  I was just
displaying bytes.len(), which is always the payload size, instead of
the actual number of message bytes.