Updates to add info to trace files
I was looking at the snake trace records and thought I saw something
strange. (It wasn’t.)  In order to make sure, I had to add the tree id
to the print for packet.  (It should have been there all along.)  When
I used that tree id, I realized my console printout of the forwarding
table didn’t have enough characters for the tree uuid.  Adding more
characters meant changing the spacing of the column headers.  A lot of
work for a small change.

I also noticed that the rust-analyzer in the web server project was
complaining about missing features.  I have no idea why it wants them,
but I added them to the Cargo.toml.  We’ll see if that resolves the
issue.