Started work on having application specify that packets not be lost on node failure
The application decides if a message is sent normally or by AIT.  I had
code in the cell agent for this, but I was setting it to false.  This
commit adds specifying AIT in the application to cell agent messages.
I also added another boolean, whether or not packets may be lost on a
node failure.  We call this Snake, because we keep a copy of the packet
on every cell it traverses.  When the packet reaches its destination,
we send a Snake ack to tell cells they can drop their copies.