Snake and SnakeD are working for leafward messages
This version works only for leafward messages.

When a packet marked Snake is received, the cmodel keeps a copy,
remembers what port it was received on, and a count of how many ports
it was sent on.  When it receives a SnakeD for that packet, it
decrements the count.  When the count reaches 0, the cmodel forgets
that packet and sends a SnakeD on the port it received the packet on.

Not implemented but worth remembering:  When the sending cell has
forgotten all the packets in a message, it can inform the sending
application that the message has been delivered to all destinations.