Make tracing infallible (in the Rust sense)
The add_to_trace() method in dal.rs returned a Result.  If people
weren’t careful to eat the error, a problem in tracing could bring the
whole program down.  This commit handles all tracing errors, such as
being unable to open a file to hold the trace records, in
add_to_trace().