Make handling of CONFIG.continue_on_error consistent
We want a panic in a thread to cause a failure in testing but not in
production.  CONFIG.continue_on_error allows us to say which behavior
we want.  I had been handling this feature differently in different
parts of the code.  This commit makes that handling uniform.

NB: This commit does not include port.rs, because someone else is
making major changes to that module.