Moved visualization web server to actix-web 2.0
When I originally coded the visualization, actix-web was at version 1.0
and did not support async.  The server doesn’t really need to be async,
but I moved the actix-web client to 2.0 in cases I want to use async in
the simulator.  Since I did that, it made sense to do the same for the
server.