A failed experiment
THIS CODE DOES NOT PRODUCE CORRECT RESULTS

The changes in this code were intended to move ENTL handling from the
packet engine to the port or even link while being compatible with
implementations for real hardware.  That change entailed additional
signaling between Port and Link and Port and PE.  I thought I had put
that signaling in, but application messages were not being forwarded.
It turns out that I would have had to poll on the signals to make
things work.

One example where I did poll is that I did circulate ENTL packets,
unlike the waiting done in what’s currently in userspace-master.  A
small example with 13 links consumes over 1/2 a CPU, even with a 5 ms
sleep on the link, while taking roughly 30 seconds to finish discovery.
 Our demo example has over 200 links.  Clearly, having the simulator
poll is not an option.

I am committing this code as the last commit on the ENTL branch because
some of the changes are worth keeping in isolation.  It’s the overall
design that’s wrong.