Added changes I want to keep from failed experiment on branch entl
I made a number of changes during the course of coding the failed
experiment, some of which are worth keeping.  These changes include a
couple of bug fixes, renaming of variables and functions to make the
code more understandable, and cleaning up some trace records.

I made one substantive change to port.rs, changing is_connected from
Arc<AtomicBool> to bool.  I don’t recall why it was an atomic, but the
code compiles and runs just fine with bool.