Prepare for new design
In the original design, the packet engine kept information needed to
recover from a link failure.  That was possible because the packet
engine decided when a packet should be sent; the port just forwarded
it.  In the new design, the port manages its own out buffer.  That
means it has the needed information, not the packet engine.

This change creates a struct FailoverInfo that the port keeps up to
date and forwards to the packet engine when a link is declared broken.
I’ve also streamlined the saved information.  It used to be able to
handle packet pipelining on the link.  It now assumes only a single
packet at a time can be in flight.

One other change was to pass in the port ID when creating the simulated
port.  Right now I’m only using it to get better trace info, but I
wonder if simulated port should have a reference to port in case it
needs to access some of port’s methods.