Simplified FailoverInfo
I realized that I don’t need a boolean to tell me if I’ve received a
packet from the other side of the link.  If packet_opt is None, then
the last packet this port sent reached the other side.  If packet_opt
is Some, then the last packet sent might have been lost.

The test for a lost packet is easy, top.  The failover message will
have a boolean indicating if the sending side has a Some or a None for
its packet_opt.  A None tells the recipient that the sender of the
failover message got the last packet by the recipient.