Fix replay to allow more trace records types during replay than during original run
We want tracing to add a minimum of overhead, so in production we only
trace what we need for visualization.  If you capture the trace of a
bad run, replay lets you run that sequence of events multiple times.
You’ll get the same trees and order of messages.  In theory, you could
also turn on more tracing, but most of those records weren’t appearing
because the messages that triggered them weren’t being sent during
replay.  This commit fixes that problem.

This commit also adds the cell ID to trace records generated in
applications.  Before the change, those trace records were captured in
a separate file.  Now they appear in the same file as the trace records
generated by the EC components, such as the cell agent and packet
engine.