Added new API to ec_message_formats.rs
The first step in implementing the new interface between port and PE is
to provide the interfaces.  This commit adds the straw man interface to
the EC message formats.  I refactored the old message types with a
suffix “Old” in order to keep the simulator running as long as possible
while implementing the new design.

I currently have placeholders for Activity, OutbufType, and the
selector for OutbufType::HeadOfLine.