Added channels for new API in parallel with old ones
I would like to be able to keep the simulator running while I code up
the new API.  This commit takes care of that part.  It adds a set of
channels between the port and PE in addition to the ones for the old
API.

No new API messages are sent, but I can now start adding them
incrementally and still run the simulator.