Code for ActivityData and moved FailoverInfo from simulated port to PE
As the first step in implementing the new API, I coded methods for
handling ActivityData.  I also moved FailoverInfo from
simulated_port.rs to port.rs.  FailoverInfo is something that all kinds
of ports will need to report, so it makes sense to define in a place
all implementations will talk to.