Added FailoverInfo to PortDisconnected status message
Since the port now handles saving the last packet sent, it has to send
that info to the packet engine when a link disconnects.  I’ve added
FailoverInfo to that message.

It looks like a lot of changes, but most of that is just renaming
PortStatus to PortStatusOld so I can use PortStatus for the new API.

The packet engine will also need the port’s out buffer(s) once a new
path has been found.  I decided to defer that step until the new path
has been found, so the packet engine can start the search just a little
bit sooner.