Added code for multiple outbuf types
Packets for control messages need to be sent at higher priority than regular packets.  I'm copying the VLanes approach from InfiniBand and having two out buffers per port, one for control messages and one for the rest.

The setup includes the ability to have additional buffers that can be used to deal with head of line blocking.  The idea is that if a port on one cell is blocked, it can tell the neighboring cell not to sent any messages on trees that go out on the blocked port.  I haven't settled on the data structure to use for that, so I have a Vec<TreeID> for now.