Handle in buffer and PE's out buffer
Added code to manage the PE's out buffer, which holds the port numbers of the ports with packets to send to the out buffer's port.  If the (currently uncoded) number of free slots in the port's out buffer is > 0, then sends the packet in the PE's in buffer and decrements the in buffer count.  If the in buffer count reaches 0, the PE sends a PeToPortPacket::Ready message to the port.  Once the port gets this Ready message, it will accept the next data packet.