Fixed a bug with control messages
The handling of control messages was working by accident, but that accident wouldn't work with the new API.  My test was if the packet was on the control tree. That test doesn't work for packets coming in on a port, because the packet has the sender's control tree ID.  The fortunate accident was that packets with unknown tree IDs get sent to the cell agent anyway.

The new API has two packet queues, one for control messages so they can have higher priority.  The old test wouldn't have worked.

This version adds a boolean is_control to the packet header.  Implementing the feature in uuid_ec.rs showed another bug where I was overwriting the flags when setting certain codes.