Simplified code setting up NOC-Port channels
There were some conditionals in datacenter.rs that I found hard to parse.  The intent was to keep any channels that might already have been defined when adding new ones.  I replaced those conditionals with hash_map::entry, which was designed for exactly this purpose.

I also added an error return if a 0 is a border port because it shows an error in the blueprint.  The code had been skipping that case.