Changed blueprint.rs to return trait Cell on get_cell()
There was an unnecessary match statement in rack.rs because blueprint.rs get_cell() returned an Either of InteriorCell or BorderCell even though there's a trait Cell.  Since rack.rs was only calling get_interior_ports() on the cell, it makes sense to use the trait.

I also made some changes in simulator.rs to be more consistent in integer type changes.  Instead of using "as" when going from a big type to a smaller one, it's safer to use try_from().

There are also some changes in utility.rs that I swear I didn't make, at least on purpose.  It looks like I accidentally ran rustfmt on the file.