Simplified port-link channels in rack.rs
Rust has a simpler way to deal with a hash map that may not have been initialized yet than what was in the code.  I used hash_map::Entry to replace 50+ lines of code with 12.  Plus the code is now Rustier.

I also changed blueprint.rs so that port 0 can never be assigned as a border or interior port.