Got 6-step ENTT working
In the olden days, an AIT transfer was a total of 6 steps, TECK-TACK-TUCK-TYCK followed by TICK-TOCK before the next transfer.  We noted that process could be shortened to TECK-TACK-TICK-TOCK.  That may have been a premature optimization, since the sender of the token got the next chance to send another one, starving the receiver.  Dale's state machine goes back to the 6-step version, which this commit follows.

This commit also includes changes to the web server's Cargo.toml to speed up compilation on a Mac as was done for the cell agent in an earlier commit.