Added option to turn off tracing of TICK/TOCK
The trace files are helpful for debugging, but turning on options that generate lots of records, such as Config.port, slows things down a LOT.  A good part of that slowdown comes from tracing TICK/TOCK packets, which are likely to be of no interest.  This commit allows you to turn on Config.port without generating trace records for TICK and TOCK by turning off Config.entl.