Added ActivityData to BasePort
Link failure can be detected either by a LINKSTATUSCHANGE event when the cable is unplugged or by a port being idle when other ports on the same cell continue to receive events.  ActivityData is used for the latter.

Every time a cell receives an event, including ENTL, it updates its ActivityData.  This data will eventually be sent to the PE, which can then decide if the link is stuck.  For now, I'm just collecting the data in BasePort and updating it in listen_link of simulated_interior_port.rs.