Some minor changes not related to new API
I found some minor issues in code for the old API when starting the code for the new API.  One is actually important.

The packet engine keeps a copy of each packet sent to the port  until it hears back from the port that the packet was received by the other side.  I was saving that packet in the wrong place, which meant it was saved even when the packet was sent to the cmodel.  The result could have been a duplicate packet following failover.