Moved listen_link_and_pe into simulated_interior_port.rs
The simulated interior port currently listens to the link and PE on different threads.  That works fine until the two sides need to access the same data.  They don't in the old API, but the new API has activity data that is shared and may have more in the future.  I could fix that problem by wrapping the shared data in Arc<Mutex<_>>, but it's easier to just use select and do everything in one thread.  Using select will also make it easier to work with both the old and new APIs  at the same time as I did in packet_engine.rs.