Realized I didn't have to tunnel messages
The cell agent listens to the NOC in a thread, the cmodel in another, and each application in its own.  Those messages affect state, but using synchronization primitives is too complicated and probably has significant performance impact.

Instead, I chose to send messages from the NOC and application threads to the cmodel thread by "bouncing" them off the cmodel.  It never occurred to me to set up a channel to send the messages directly from one cell agent thread to the cell agent's cmodel thread.  That's what this version does by using the cm_to_ca channel solely in the cell agent.

The CmToCaBytes enum still has entries for those messages, but they are never used.