Fixed a lost handshake
When I moved ENTT handling from the packet engine into the port, I lost a handshake that kept the port from interleaving ENTT packets on its link.

In this commit, the port's thread that talks to the PE blocks until the packet transfer is complete.  That means blocking until the port has sent a Normal packet or received a TUCK or TYCK.