Some changes to cross compile to Linux
I was hoping that I would be able to link the cell version of the code if I cross compiled to Linux.  That didn't work because I can't compile the C code in the ecnl directory without the NetLink headers, and I don't know which ones to download.  I'm keeping these changes since they do no harm and might prove useful later